requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2.6
django-cors-headers==3.8.0
django-crispy-forms==1.13.0
djangorestframework==3.12.4
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
wincertstore==0.2
